start moving ScopeGraph to its own project
